Fix case where --socket was not always been stored correctly.

+++ libmemcached/backtrace.cc
@@ -59,7 +59,7 @@ void custom_backtrace(void)
 #ifdef HAVE_EXECINFO_H
   void *array[50];
 
-  size_t size= backtrace(array, 50);
+  int size= backtrace(array, 50);
   char **strings= backtrace_symbols(array, size);
 
   if (strings == NULL)
@@ -67,7 +67,7 @@ void custom_backtrace(void)
     return;
   }
 
-  fprintf(stderr, "Number of stack frames obtained: %lu\n", (unsigned long)size);
+  fprintf(stderr, "Number of stack frames obtained: %d\n", size);
 
   char *named_function= (char *)std::realloc(NULL, 1024);
   
@@ -77,7 +77,7 @@ void custom_backtrace(void)
     return;
   }
 
-  for (size_t x= 1; x < size; x++) 
+  for (int x= 1; x < size; x++) 
   {
     if (USE_DEMANGLE)
     {
